fix auth admin and user

+++ routes/web.php
@@ -3,6 +3,7 @@
 use App\Http\Controllers\AdminController;
 use App\Http\Controllers\ProductController;
 use App\Http\Controllers\UserController;
+use App\Http\Middleware\Authenticate;
 use Illuminate\Support\Facades\Route;
 
 /*
@@ -16,20 +17,21 @@ use Illuminate\Support\Facades\Route;
 |
 */
 
-Route::get('/', [UserController::class, 'landingPageView']);
 
-Route::get('/login', [UserController::class, 'loginView'])->name('login');
+Route::get('/login', [UserController::class, 'loginView'])->name('login')->middleware('guest');
 Route::post('/login', [UserController::class, 'loginUser']);
 
-Route::get('/register', [UserController::class, 'registerView'])->name('register');
+Route::get('/register', [UserController::class, 'registerView'])->middleware('guest')->name('register');
 Route::post('/register', [UserController::class, 'registerUser']);
 
-Route::post('/logout', [UserController::class, 'logoutUser']);
+Route::post('/logout', [UserController::class, 'logoutUser'])->name('logout');
+
+Route::get('/', [UserController::class, 'landingPageView']);
 
 Route::prefix('/admin')->group(function () {
-    Route::get('/', [AdminController::class, 'dashboardView'])->name('dashboard');
-    Route::get('/listProduct', [AdminController::class, 'showProducts'])->name('listProduct');
-    Route::get('/addProduct', [AdminController::class, 'addProductView'])->name('addProduct');
-    Route::post('/saveProduct', [ProductController::class, 'saveProduct'])->name('saveProduct');
-    Route::delete('/deleteProduct/{id}', [ProductController::class, 'deleteProduct'])->name('deleteProduct');
+    Route::get('/listProduct', [AdminController::class, 'showProducts'])->middleware('auth')->name('listProduct');
+    Route::get('/addProduct', [AdminController::class, 'addProductView'])->middleware('auth')->name('addProduct');
+    Route::post('/saveProduct', [ProductController::class, 'saveProduct'])->middleware('auth')->name('saveProduct');
+    Route::delete('/deleteProduct/{id}', [ProductController::class, 'deleteProduct'])->middleware('auth')->name('deleteProduct');
+    Route::get('/', [AdminController::class, 'dashboardView'])->name('dashboard')->middleware('auth');
 });
